requirements.txt
asgiref==3.4.1
Django==3.2.9
django-polls==0.1
django-taggit==1.5.1
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
